Made mailgun_rails depend on actionmailer, not rails
Previously, the gem was pulling in the rails gem as a dependency,
meaning that any project using mailgun_rails have all of the gems in
rails as dependencies too. That's an issue if you are not making a rails
app, or just want to keep your rails app as lean as possible.
mailgun_rails works fine with just actionmailer as a dependency, so I've
swapped rails for actionmailer and json as dependencies instead.