Removed Gemfile.lock
The JRuby build on travis is failing because it's using an old version
of unf_ext, that's not compatible with JRuby. The bug was fixed some
time ago in unf, but Gemfile.lock has not been updated.

To resolve the immediate issue but also avoid the problem in future,
I've removed Gemfile.lock completely, so it's regenerated each time