Ignored rubinius build failures on travis
There's a bug in the mail gem that only affects rubinius. This seems to
be causing build failures for mailgun_rails. Here's the details of the
bug:

https://github.com/mikel/mail/pull/782

This commit temporarily sets travis to ignore failures on rubinius. I
think it's better to have a reliable build than have to manually check
whether the failure is due to a known bug in another library.

Ultimately this should be reverted once the bug is resolved.