Pass on any specific error message returned from Mailgun
When the Mailgun API returns a "400 Bad Request" error, it includes the specific error message (for
example, "'to' parameter is missing") from in the JSON response.

However, there was no easy way to see what that specific error message actually was, and you'd end
up with a generic (and therefore useless) RestClient::BadRequest error from your Rails app that you
had no idea the actual cause of.

This change causes it to automatically extract that specific message from the JSON response and use
that as the Ruby exception message.